fix(): change credit option to float type

+++ src/main/java/com/dell/vendingmachine/model/CreditOption.java
@@ -6,15 +6,12 @@ import javax.persistence.*;
 import java.io.Serializable;
 
 @Entity
-public @Data @Builder @NoArgsConstructor
-@AllArgsConstructor
-class CreditOption  implements Serializable {
+public @Data @Builder @NoArgsConstructor @AllArgsConstructor class CreditOption implements Serializable {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long creditOptionId;
 
-    private  String label;
-
-    private double value;
+    private String label;
+    private float value;
 }
